+++ modules/instructions/handle_async_operations.go
@@ -1,4 +1,4 @@
-package messages
+package instructions
 
 import (
 	"time"
